// Fixr/src/app/models/enums/statusAnuncio.enum.ts
export enum StatusAnuncio {
    RASCUNHO = 'RASCUNHO',
    PUBLICADO = 'PUBLICADO',
    PAUSADO = 'PAUSADO',
    EM_NEGOCIACAO = 'EM_NEGOCIACAO',
    ACORDO_FECHADO = 'ACORDO_FECHADO',
    EM_ANDAMENTO = 'EM_ANDAMENTO',
    CONCLUIDO = 'CONCLUIDO',
    CANCELADO = 'CANCELADO',
    EXPIRADO = 'EXPIRADO',
    ARQUIVADO = 'ARQUIVADO'
}